Fix and update dashboard page and other UI

+++ src/components/DeleteAlert.tsx
@@ -1,12 +1,5 @@
-'use cient'
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card"
+'use client';
+
 import {
   AlertDialog,
   AlertDialogAction,
@@ -19,27 +12,30 @@ import {
   AlertDialogTrigger,
 } from "@/components/ui/alert-dialog"
 import { Button } from "./ui/button"
-import { X } from "lucide-react"
+import { MoreHorizontal, Trash2} from "lucide-react"
 import { Message } from "@/model/User"
 import { useToast } from "./ui/use-toast"
 import axios, { AxiosError } from "axios"
 import { ApiResponse } from "@/types/ApiResponse"
 
-type MessageCardProp = {
-  message: Message
+type DeleteAlertProp = {
+  message?: Message
+  messages?: Message[]
   onMessageDelete: (messageId: string) => void
 }
 
-export function Alert({ message, onMessageDelete }: MessageCardProp) {
+export function Alert({ message, onMessageDelete }: DeleteAlertProp) {
   const { toast } = useToast()
   const handleDeleteConfirm = async () => {
     try {
-      const response = await axios.delete<ApiResponse>(`/api/delete-message/${message._id}`)
-      toast({
-        title: response.data.message
-      })
-      const messageId = message._id as string
-      onMessageDelete(messageId)
+      if (message) {
+        const response = await axios.delete<ApiResponse>(`/api/delete-message/${message._id}`)
+        toast({
+          title: response.data.message
+        })
+        const messageId = message._id as string
+        onMessageDelete(messageId)
+      }
     } catch (error) {
       const axiosError = error as AxiosError<ApiResponse>;
       toast({
@@ -54,14 +50,15 @@ export function Alert({ message, onMessageDelete }: MessageCardProp) {
   return (
     <AlertDialog>
       <AlertDialogTrigger asChild>
-        <Button variant="destructive"><X className="w-5 h-5" /></Button>
+        <Button variant="ghost" className="p-0 w-full">
+           Delete
+        </Button>
       </AlertDialogTrigger>
       <AlertDialogContent>
         <AlertDialogHeader>
           <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
           <AlertDialogDescription>
-            This action cannot be undone. This will permanently delete your
-            account and remove your data from our servers.
+            This action cannot be undone. This will permanently delete the message from our servers.
           </AlertDialogDescription>
         </AlertDialogHeader>
         <AlertDialogFooter>
@@ -73,33 +70,3 @@ export function Alert({ message, onMessageDelete }: MessageCardProp) {
   )
 }
 
-
-
-function MessageCard({ message, onMessageDelete }: MessageCardProp) {
-  const dateString = new Date(message.createdAt)
-    .toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-      hour: 'numeric',
-      minute: 'numeric',
-      second: 'numeric',
-      hour12: true
-    });
-  return (
-    <Card>
-      <CardHeader>
-        <div className="flex justify-between items-center">
-          <CardTitle>{message.content}</CardTitle>
-          <Alert message={message} onMessageDelete={onMessageDelete} />
-        </div>
-
-        <div className="text-sm">
-          {dateString}
-        </div>
-      </CardHeader>
-    </Card>
-  )
-}
-
-export default MessageCard
